+++ src/main/java/kr/hhplus/be/server/presentation/KafkaPublishingException.java
@@ -1,4 +1,4 @@
-package kr.hhplus.be.server.presentation.payment;
+package kr.hhplus.be.server.presentation;
 
 public class KafkaPublishingException extends Throwable {
     public KafkaPublishingException(String s, Exception e) {
